Use automake's parallel tests feature.
This simplifies the various clamd/clamscan tests a lot,
since the valgrind tests can now be parallelized by automake,
and we don't need to do that in the script itself.

+++ unit_tests/valgrind_tests.sh
@@ -17,12 +17,13 @@ parse_valgrindlog()
 	if test ! -f $1; then
 		echo "*** Logfile $1 not found. Valgrind failed to run?"
 	fi
-	NRUNS=`grep "ERROR SUMMARY" $1 | wc -l`
-	if test $NRUNS -eq `grep "ERROR SUMMARY: 0 errors" $1 | wc -l` && test `grep "FATAL:" $1|wc -l ` -eq 0; then
+	NRUNS=`grep -a "ERROR SUMMARY" $1 | wc -l`
+	if test $NRUNS -eq `grep -a "ERROR SUMMARY: 0 errors" $1 | wc -l` && test `grep -a "FATAL:" $1|wc -l ` -eq 0; then
 		if test "$1" = "valgrind-race.log" || 
 			test $NRUNS -eq `grep "no leaks are possible" $1 | wc -l` ||
 			test `grep "lost:" $1 | grep -v " 0 bytes" | wc -l` -eq 0; then 
 			if test -z "$GENSUPP"; then
+			        cp $1 /tmp
 				rm -f $1;
 			fi
 			return
@@ -62,24 +63,40 @@ else
 	GENSUPP=
 fi
 
-VALGRIND_COMMON_FLAGS="-v --trace-children=yes --suppressions=$abs_srcdir/valgrind.supp $GENSUPP"
+VALGRIND_COMMON_FLAGS="-v --trace-children=yes --suppressions=$abs_srcdir/valgrind.supp --error-exitcode=123 $GENSUPP"
 VALGRIND_FLAGS="$VALGRIND_COMMON_FLAGS --track-fds=yes --leak-check=full"
 VALGRIND_FLAGS_RACE="$VALGRIND_COMMON_FLAGS --tool=helgrind"
 export CK_DEFAULT_TIMEOUT=40
 echo "--- Starting check_clamav under valgrind/memcheck"
 rm -f valgrind-check.log valgrind-clamd.log valgrind-race.log
-CK_FORK=no ../libtool --mode=execute $VALGRIND $VALGRIND_FLAGS ./check_clamav >valgrind-check.log 2>&1 &
+CK_FORK=no ../libtool --mode=execute $VALGRIND $VALGRIND_FLAGS --log-file=valgrind-check.log ./check_clamav &
 pid1=$!
 
 echo "--- Starting clamd under valgrind/memcheck"
-CLAMD_WRAPPER="$VALGRIND $VALGRIND_FLAGS" $abs_srcdir/check_clamd.sh >valgrind-clamd.log 2>&1 &
+CLAMD_WRAPPER="$VALGRIND $VALGRIND_FLAGS --log-file=`pwd`/valgrind-clamd.log" $abs_srcdir/check_clamd.sh &
 pid2=$!
 
 echo "--- Starting clamd under valgrind/helgrind"
-CLAMD_TEST_UNIQ1=3 CLAMD_TEST_UNIQ2=4 CLAMD_WRAPPER="$VALGRIND $VALGRIND_FLAGS_RACE" $abs_srcdir/check_clamd.sh >valgrind-race.log 2>&1 &
+CLAMD_TEST_UNIQ1=3 CLAMD_TEST_UNIQ2=4 CLAMD_WRAPPER="$VALGRIND $VALGRIND_FLAGS_RACE --log-file=`pwd`/valgrind-race.log" $abs_srcdir/check_clamd.sh &
 pid3=$!
 
-wait $pid1 $pid2 $pid3
+errors=0
+wait $pid1
+if test $? -ne 0; then
+    # don't rm logs
+    GENSUPP=1
+    errors=1
+fi
+wait $pid2
+if test $? -ne 0; then
+    GENSUPP=1
+    errors=1
+fi
+wait $pid3
+if test $? -ne 0; then
+    GENSUPP=1
+    errors=1
+fi
 parse_valgrindlog valgrind-check.log
 parse_valgrindlog valgrind-clamd.log
 parse_valgrindlog valgrind-race.log
@@ -87,4 +104,4 @@ parse_valgrindlog valgrind-race.log
 if test -f valgrind-check.log || test -f valgrind-race.log || test -f valgrind-clamd.log; then
 	exit 1;
 fi
-exit 0
+exit $errors
